Add print.hpp and move printing code there.
